Fix several warnings
* Avoid many implicit type conversions or make them at least explicit
    * Use range-based for loops instead of an explicit index where possible
      to eliminate the problematic variable/comparison completely
    * Use float literals to avoid implicit conversions to double (and then
      back to float)
    * Replace some macros with constexpr functions that return the correct
      type
* Port away from deprecated FluidSynth and Qt classes and functions, most
  notably:
    * Use QElapsedTimer instead of QTime and change the type used to store
      ticks to qint64 accordingly
* Declare unused functions parameters as such explicitly to silence related
  warnings

+++ src/TrackList.cpp
@@ -366,7 +366,7 @@ int CTrackList::getHandTrackIndex(whichPart_t whichPart)
     return -1;
 }
 
-void CTrackList::changeListWidgetItemView( unsigned int index, QListWidgetItem* listWidgetItem )
+void CTrackList::changeListWidgetItemView( int index, QListWidgetItem* listWidgetItem )
 {
     int chan = m_partsList[index].midiChannel();
     if ( CNote::hasPianoPart( chan ))
@@ -534,7 +534,7 @@ QString CTrackList::getProgramName(int program)
         /* 128. */ QT_TR_NOOP("Gunshot"),
     };
 
-    if (program >= 0 && program < static_cast<int>(arraySize(gmInstrumentNames)))
+    if (program >= 0 && program < arraySize(gmInstrumentNames))
         return tr(gmInstrumentNames[program]);
     else
         return QString();
